+++ docs/src/main/paradox/release-notes.md
@@ -6,7 +6,7 @@
 
 * Added `method_name` parameter to the `rf_resample` method.
    * __BREAKING__: In SQL, the function `rf_resample` now takes 3 arguments. You can use `rf_resample_nearest` with two arguments or refactor to `rf_resample(t, v, "nearest")`.
-* Added resample method parameter to SQL and Python APIs. This will affect the reprojection of right hand side tiles.
+* Added resample method parameter to SQL and Python APIs. @ref:[See updated docs](raster-join.md).
 
 
 ### 0.9.0